india_art_culture_2/requirements.txt
streamlit==1.32.0
snowflake-connector-python==3.7.0
pandas==2.2.0
plotly==5.18.0
numpy==2.2.6
python-dotenv==1.0.0
folium==0.15.1
streamlit-folium==0.18.0
seaborn>=0.13.0
matplotlib>=3.8.0
requests==2.31.0
beautifulsoup4==4.12.0
pillow==10.2.0
cryptography>=42.0.5 